디렉토리명 변경, 최초 점속 account web3 연동
