+++ app_client/package.json
@@ -3,11 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "ethereumjs-testrpc": "^6.0.3",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "react-scripts": "1.1.4",
-    "truffle-contract": "^4.0.0-next.0",
-    "web3": "^1.0.0-beta.35"
+    "truffle": "^4.1.14",
+    "truffle-hdwallet-provider": "0.0.6",
+    "web3": "^0.20.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -19,6 +21,7 @@
     "link-contracts:win32": "cd src && mklink \\D contracts ..\\..\\build\\contracts"
   },
   "devDependencies": {
-    "run-script-os": "^1.0.3"
+    "run-script-os": "^1.0.3",
+    "web3": "^1.0.0-beta.36"
   }
 }
