+++ app_client/public/manifest.json
