+++ app_client/src/utils/getWeb3.js
@@ -17,7 +17,7 @@ const getWeb3 = () =>
       } else {
         // Fallback to localhost if no web3 injection. We've configured this to
         // use the development console's port by default.
-        const provider = new Web3.providers.HttpProvider(
+        const provider = new Web3.providers.HDWalletProvider(
           "http://127.0.0.1:9545"
         );
         web3 = new Web3(provider);
